requirements.txt
speedtest-cli
dash
pandas
python-dotenv
plotly
dash-core-components
dash-html-components
dash-table
gunicorn 